Update yaac-another-awesome-cv.cls
Hi, I am using this template for my resume, I like it! I did a minor change. I changed the command for \sociallink and \socialtext. I noticed the text wasn't lined out after the symbols, if compared to each other. I used a tabular now. To me this is more neat. It can be a matter of taste though.